Refactored application for dependency injection.
Updated the application to use the server resources `RootServerResource` and `HelloWorldServerResource`. This included refactoring the `HelloWorldResource` and `RootResource` classes into interfaces. As part of this change, the `Main.java` file was adjusted to work with the new server resource and module package implementations. Supplementary changes were made to `pom.xml` for packaging uber jar, updated `Configuration.java` to be usable within the inbound route without DI.